src/requirements.txt
acme==1.21.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
ansible-base==2.10.8
apache-libcloud==3.2.0
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.8.1
attrs==23.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certbot==1.21.0
certifi==2024.7.4
cffi==1.17.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cx_Freeze==7.2.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
evdev==1.4.0
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.15.4
fonttools==4.53.1
future==0.18.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
josepy==1.10.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
macholib==1.16.3
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.2
modulegraph==0.19.6
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
numpy==2.0.1
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.1
paramiko==2.9.3
parsedatetime==2.6
patchelf==0.17.2.1
pexpect==4.8.0
pillow==10.4.0
pluggy==1.5.0
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
py2app==0.28.8
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==8.2.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selenium==4.22.0
selinux==3.3
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.5.1
ssh-import-id==5.11
sympy==1.13.2
systemd-python==234
termcolor==2.4.0
texttable==1.6.4
tk==0.1.0
toml==0.10.2
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
WoeUSB-ng==0.2.12
wsproto==1.2.0
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
